requirements.txt
python-telegram-bot==21.9
aiosqlite==0.20.0
APScheduler==3.10.4
httpx==0.27.0
timezonefinder==6.5.2
pytz==2024.1
